+++ src/domain/culture_base.cpp
@@ -714,7 +714,7 @@ void culture_base::add_names_from(const culture_base *other)
 	}
 }
 
-const phenotype_map<int> &culture_base::get_phenotype_weights() const
+const phenotype_map<int64_t> &culture_base::get_phenotype_weights() const
 {
 	if (!this->phenotype_weights.empty()) {
 		return this->phenotype_weights;
@@ -727,9 +727,9 @@ const phenotype_map<int> &culture_base::get_phenotype_weights() const
 	return this->phenotype_weights;
 }
 
-void culture_base::change_phenotype_weight(const phenotype *phenotype, const int change)
+void culture_base::change_phenotype_weight(const phenotype *phenotype, const int64_t change)
 {
-	const int weight = (this->phenotype_weights[phenotype] += change);
+	const int64_t weight = (this->phenotype_weights[phenotype] += change);
 
 	if (weight == 0) {
 		this->phenotype_weights.erase(phenotype);
